Update DrawEngine/windoge_make.rb to be more intelligent
I made a module wrapper in that script. This is good because I can
insert RPG Maker styled comment blocks then.

The script can now output a batch file if it's not been generated
before and after that, you can double-click on the batch file directly
to compile the Ruby extension.